Refactored useCheckExisting hook to suggestions
