+++ src/lib/hooks/useCheckExisting.js
@@ -1,6 +1,6 @@
 import { useCallback, useState } from 'react'
 
-import { useAlreadySignedUp } from '../../components/auth/torus/AuthTorus'
+import { useAlreadySignedUp } from '../../components/auth/torus/hooks/useAlreadySignedUp.js'
 import logger from '../../lib/logger/js-logger'
 import { userExists } from '../../lib/login/userExists'
 
@@ -8,21 +8,21 @@ const log = logger.child({ from: 'useCheckExisting' })
 
 //check if email/mobile was used to register before and offer user to login instead
 const useCheckExisting = (torusProvider, navigation) => {
-  const [selection, setSelection] = useState('signup')
+  const [, setSelection] = useState('signup')
   const showAlreadySignedUp = useAlreadySignedUp()
 
   const checkExisting = useCallback(
     async searchBy => {
-      const existsResult = await userExists(searchBy).catch(e => {
+      const { exists } = (await userExists(searchBy).catch(e => {
         log.warn('userExists check failed:', e.message, e)
-        return { exists: false }
-      })
+      })) || { exists: false }
 
-      log.debug('checking userAlreadyExist', { existsResult })
+      log.debug('checking userAlreadyExist', { exists })
 
-      if (existsResult.exists) {
-        setSelection(await showAlreadySignedUp(torusProvider, existsResult, searchBy.email ? 'email' : 'mobile'))
-        if (selection === 'signin') {
+      if (exists) {
+        const nextStep = await showAlreadySignedUp(torusProvider, exists, searchBy.email ? 'email' : 'mobile')
+        setSelection(nextStep)
+        if (nextStep === 'signin') {
           return navigation.navigate('Auth', { screen: 'signin' })
         }
       }
